Bug1875563-add profile caAuditSigningCert
This patch will revert the previously added IPA specific KRA storage
and transport cert prorfiles, as it turned out that they just need
generic KRA storage and transport cert profiles, which could be
fulfilled by using the following two existing profiles
  caStorageCert.cfg
  caTransportCert.cfg
In addition, a caAuditSigningCert profile is added, although I find
a misleading profile named caSignedLogCert.cfg  that was intended for
the use.  I disabled caSignedLogCert.cfg instead.

I also removed the SHA1 algorithms from all the *storage* and *audit*
profiles while I'm at it.

The upgrade scripts only adds the new profile caAuditSigningCert.  It
does not modify existing profiles or remove those two IPA specific
ones.

fixes https://bugzilla.redhat.com/show_bug.cgi?id=1875563
