+++ base/ca/shared/profiles/ca/caAuditSigningCert.cfg
@@ -1,10 +1,10 @@
 desc=This certificate profile is for enrolling audit signing certificates.
-visible=false
+visible=true
 enable=true
 enableBy=admin
-auth.instance_id=TokenAuth
+auth.instance_id=
 authz.acl=group="Enterprise OCSP Administrators" || group="Enterprise RA Administrators" || group="Enterprise CA Administrators" || group="Enterprise KRA Administrators" || group="Enterprise TKS Administrators" || group="Enterprise TPS Administrators"
-name=Audit Signing Certificate Enrollment
+name=Manual Audit Signing Certificate Enrollment
 input.list=i1,i2
 input.i1.class_id=certReqInputImpl
 input.i2.class_id=submitterInfoInputImpl
