--- base/ca/shared/profiles/ca/caIPAKraStorageCert.cfg
@@ -1,79 +1,0 @@
-desc=This certificate profile is for enrolling IPA KRA storage certificates.
-visible=false
-enable=true
-enableBy=admin
-auth.instance_id=raCertAuth
-name=IPA-RA Agent-Authenticated KRA Storage Certificate Enrollment
-input.list=i1,i2
-input.i1.class_id=certReqInputImpl
-input.i2.class_id=submitterInfoInputImpl
-output.list=o1
-output.o1.class_id=certOutputImpl
-policyset.list=drmStorageCertSet
-policyset.drmStorageCertSet.list=1,2,3,4,5,6,9
-policyset.drmStorageCertSet.1.constraint.class_id=subjectNameConstraintImpl
-policyset.drmStorageCertSet.1.constraint.name=Subject Name Constraint
-policyset.drmStorageCertSet.1.constraint.params.pattern=CN=.*
-policyset.drmStorageCertSet.1.constraint.params.accept=true
-policyset.drmStorageCertSet.1.default.class_id=userSubjectNameDefaultImpl
-policyset.drmStorageCertSet.1.default.name=Subject Name Default
-policyset.drmStorageCertSet.1.default.params.name=
-policyset.drmStorageCertSet.2.constraint.class_id=validityConstraintImpl
-policyset.drmStorageCertSet.2.constraint.name=Validity Constraint
-policyset.drmStorageCertSet.2.constraint.params.range=720
-policyset.drmStorageCertSet.2.constraint.params.notBeforeCheck=false
-policyset.drmStorageCertSet.2.constraint.params.notAfterCheck=false
-policyset.drmStorageCertSet.2.default.class_id=validityDefaultImpl
-policyset.drmStorageCertSet.2.default.name=Validity Default
-policyset.drmStorageCertSet.2.default.params.range=720
-policyset.drmStorageCertSet.2.default.params.startTime=0
-policyset.drmStorageCertSet.3.constraint.class_id=keyConstraintImpl
-policyset.drmStorageCertSet.3.constraint.name=Key Constraint
-policyset.drmStorageCertSet.3.constraint.params.keyType=RSA
-policyset.drmStorageCertSet.3.constraint.params.keyParameters=1024,2048,3072,4096
-policyset.drmStorageCertSet.3.default.class_id=userKeyDefaultImpl
-policyset.drmStorageCertSet.3.default.name=Key Default
-policyset.drmStorageCertSet.4.constraint.class_id=noConstraintImpl
-policyset.drmStorageCertSet.4.constraint.name=No Constraint
-policyset.drmStorageCertSet.4.default.class_id=authorityKeyIdentifierExtDefaultImpl
-policyset.drmStorageCertSet.4.default.name=Authority Key Identifier Default
-policyset.drmStorageCertSet.5.constraint.class_id=noConstraintImpl
-policyset.drmStorageCertSet.5.constraint.name=No Constraint
-policyset.drmStorageCertSet.5.default.class_id=authInfoAccessExtDefaultImpl
-policyset.drmStorageCertSet.5.default.name=AIA Extension Default
-policyset.drmStorageCertSet.5.default.params.authInfoAccessADEnable_0=true
-policyset.drmStorageCertSet.5.default.params.authInfoAccessADLocationType_0=URIName
-policyset.drmStorageCertSet.5.default.params.authInfoAccessADLocation_0=
-policyset.drmStorageCertSet.5.default.params.authInfoAccessADMethod_0=1.3.6.1.5.5.7.48.1
-policyset.drmStorageCertSet.5.default.params.authInfoAccessCritical=false
-policyset.drmStorageCertSet.5.default.params.authInfoAccessNumADs=1
-policyset.drmStorageCertSet.6.constraint.class_id=keyUsageExtConstraintImpl
-policyset.drmStorageCertSet.6.constraint.name=Key Usage Extension Constraint
-policyset.drmStorageCertSet.6.constraint.params.keyUsageCritical=true
-policyset.drmStorageCertSet.6.constraint.params.keyUsageDigitalSignature=true
-policyset.drmStorageCertSet.6.constraint.params.keyUsageNonRepudiation=true
-policyset.drmStorageCertSet.6.constraint.params.keyUsageDataEncipherment=true
-policyset.drmStorageCertSet.6.constraint.params.keyUsageKeyEncipherment=true
-policyset.drmStorageCertSet.6.constraint.params.keyUsageKeyAgreement=false
-policyset.drmStorageCertSet.6.constraint.params.keyUsageKeyCertSign=false
-policyset.drmStorageCertSet.6.constraint.params.keyUsageCrlSign=false
-policyset.drmStorageCertSet.6.constraint.params.keyUsageEncipherOnly=false
-policyset.drmStorageCertSet.6.constraint.params.keyUsageDecipherOnly=false
-policyset.drmStorageCertSet.6.default.class_id=keyUsageExtDefaultImpl
-policyset.drmStorageCertSet.6.default.name=Key Usage Default
-policyset.drmStorageCertSet.6.default.params.keyUsageCritical=true
-policyset.drmStorageCertSet.6.default.params.keyUsageDigitalSignature=true
-policyset.drmStorageCertSet.6.default.params.keyUsageNonRepudiation=true
-policyset.drmStorageCertSet.6.default.params.keyUsageDataEncipherment=true
-policyset.drmStorageCertSet.6.default.params.keyUsageKeyEncipherment=true
-policyset.drmStorageCertSet.6.default.params.keyUsageKeyAgreement=false
-policyset.drmStorageCertSet.6.default.params.keyUsageKeyCertSign=false
-policyset.drmStorageCertSet.6.default.params.keyUsageCrlSign=false
-policyset.drmStorageCertSet.6.default.params.keyUsageEncipherOnly=false
-policyset.drmStorageCertSet.6.default.params.keyUsageDecipherOnly=false
-policyset.drmStorageCertSet.9.constraint.class_id=signingAlgConstraintImpl
-policyset.drmStorageCertSet.9.constraint.name=No Constraint
-policyset.drmStorageCertSet.9.constraint.params.signingAlgsAllowed=SHA1withRSA,SHA256withRSA,SHA512withRSA,SHA1withDSA,SHA1withEC,SHA256withEC,SHA384withRSA,SHA384withEC,SHA512withEC,SHA256withRSA/PSS,SHA384withRSA/PSS,SHA512withRSA/PSS
-policyset.drmStorageCertSet.9.default.class_id=signingAlgDefaultImpl
-policyset.drmStorageCertSet.9.default.name=Signing Alg
-policyset.drmStorageCertSet.9.default.params.signingAlg=-
