--- base/ca/shared/profiles/ca/caIPAKraTransportCert.cfg
@@ -1,85 +1,0 @@
-desc=This certificate profile is for enrolling IPA KRA transport certificates.
-visible=false
-enable=true
-enableBy=admin
-auth.instance_id=raCertAuth
-name=IPA-RA Agent-Authenticated KRA Transport Certificate Enrollment
-input.list=i1,i2
-input.i1.class_id=certReqInputImpl
-input.i2.class_id=submitterInfoInputImpl
-output.list=o1
-output.o1.class_id=certOutputImpl
-policyset.list=transportCertSet
-policyset.transportCertSet.list=1,2,3,4,5,6,7,8
-policyset.transportCertSet.1.constraint.class_id=subjectNameConstraintImpl
-policyset.transportCertSet.1.constraint.name=Subject Name Constraint
-policyset.transportCertSet.1.constraint.params.pattern=CN=.*
-policyset.transportCertSet.1.constraint.params.accept=true
-policyset.transportCertSet.1.default.class_id=userSubjectNameDefaultImpl
-policyset.transportCertSet.1.default.name=Subject Name Default
-policyset.transportCertSet.1.default.params.name=
-policyset.transportCertSet.2.constraint.class_id=validityConstraintImpl
-policyset.transportCertSet.2.constraint.name=Validity Constraint
-policyset.transportCertSet.2.constraint.params.range=720
-policyset.transportCertSet.2.constraint.params.notBeforeCheck=false
-policyset.transportCertSet.2.constraint.params.notAfterCheck=false
-policyset.transportCertSet.2.default.class_id=validityDefaultImpl
-policyset.transportCertSet.2.default.name=Validity Default
-policyset.transportCertSet.2.default.params.range=720
-policyset.transportCertSet.2.default.params.startTime=0
-policyset.transportCertSet.3.constraint.class_id=keyConstraintImpl
-policyset.transportCertSet.3.constraint.name=Key Constraint
-policyset.transportCertSet.3.constraint.params.keyType=RSA
-policyset.transportCertSet.3.constraint.params.keyParameters=1024,2048,3072,4096
-policyset.transportCertSet.3.default.class_id=userKeyDefaultImpl
-policyset.transportCertSet.3.default.name=Key Default
-policyset.transportCertSet.4.constraint.class_id=noConstraintImpl
-policyset.transportCertSet.4.constraint.name=No Constraint
-policyset.transportCertSet.4.default.class_id=authorityKeyIdentifierExtDefaultImpl
-policyset.transportCertSet.4.default.name=Authority Key Identifier Default
-policyset.transportCertSet.5.constraint.class_id=noConstraintImpl
-policyset.transportCertSet.5.constraint.name=No Constraint
-policyset.transportCertSet.5.default.class_id=authInfoAccessExtDefaultImpl
-policyset.transportCertSet.5.default.name=AIA Extension Default
-policyset.transportCertSet.5.default.params.authInfoAccessADEnable_0=true
-policyset.transportCertSet.5.default.params.authInfoAccessADLocationType_0=URIName
-policyset.transportCertSet.5.default.params.authInfoAccessADLocation_0=
-policyset.transportCertSet.5.default.params.authInfoAccessADMethod_0=1.3.6.1.5.5.7.48.1
-policyset.transportCertSet.5.default.params.authInfoAccessCritical=false
-policyset.transportCertSet.5.default.params.authInfoAccessNumADs=1
-policyset.transportCertSet.6.constraint.class_id=keyUsageExtConstraintImpl
-policyset.transportCertSet.6.constraint.name=Key Usage Extension Constraint
-policyset.transportCertSet.6.constraint.params.keyUsageCritical=true
-policyset.transportCertSet.6.constraint.params.keyUsageDigitalSignature=true
-policyset.transportCertSet.6.constraint.params.keyUsageNonRepudiation=true
-policyset.transportCertSet.6.constraint.params.keyUsageDataEncipherment=true
-policyset.transportCertSet.6.constraint.params.keyUsageKeyEncipherment=true
-policyset.transportCertSet.6.constraint.params.keyUsageKeyAgreement=false
-policyset.transportCertSet.6.constraint.params.keyUsageKeyCertSign=false
-policyset.transportCertSet.6.constraint.params.keyUsageCrlSign=false
-policyset.transportCertSet.6.constraint.params.keyUsageEncipherOnly=false
-policyset.transportCertSet.6.constraint.params.keyUsageDecipherOnly=false
-policyset.transportCertSet.6.default.class_id=keyUsageExtDefaultImpl
-policyset.transportCertSet.6.default.name=Key Usage Default
-policyset.transportCertSet.6.default.params.keyUsageCritical=true
-policyset.transportCertSet.6.default.params.keyUsageDigitalSignature=true
-policyset.transportCertSet.6.default.params.keyUsageNonRepudiation=true
-policyset.transportCertSet.6.default.params.keyUsageDataEncipherment=true
-policyset.transportCertSet.6.default.params.keyUsageKeyEncipherment=true
-policyset.transportCertSet.6.default.params.keyUsageKeyAgreement=false
-policyset.transportCertSet.6.default.params.keyUsageKeyCertSign=false
-policyset.transportCertSet.6.default.params.keyUsageCrlSign=false
-policyset.transportCertSet.6.default.params.keyUsageEncipherOnly=false
-policyset.transportCertSet.6.default.params.keyUsageDecipherOnly=false
-policyset.transportCertSet.7.constraint.class_id=noConstraintImpl
-policyset.transportCertSet.7.constraint.name=No Constraint
-policyset.transportCertSet.7.default.class_id=extendedKeyUsageExtDefaultImpl
-policyset.transportCertSet.7.default.name=Extended Key Usage Extension Default
-policyset.transportCertSet.7.default.params.exKeyUsageCritical=false
-policyset.transportCertSet.7.default.params.exKeyUsageOIDs=1.3.6.1.5.5.7.3.2
-policyset.transportCertSet.8.constraint.class_id=signingAlgConstraintImpl
-policyset.transportCertSet.8.constraint.name=No Constraint
-policyset.transportCertSet.8.constraint.params.signingAlgsAllowed=SHA256withRSA,SHA384withRSA,SHA512withRSA,SHA256withEC,SHA384withEC,SHA512withEC,SHA256withRSA/PSS,SHA384withRSA/PSS,SHA512withRSA/PSS
-policyset.transportCertSet.8.default.class_id=signingAlgDefaultImpl
-policyset.transportCertSet.8.default.name=Signing Alg
-policyset.transportCertSet.8.default.params.signingAlg=-
