+++ base/ca/shared/profiles/ca/caSignedLogCert.cfg
@@ -1,6 +1,6 @@
 desc=This profile is for enrolling audit log signing certificates
-visible=true
-enable=true
+visible=false
+enable=false
 enableBy=admin
 auth.class_id=
 name=Manual Log Signing Certificate Enrollment
@@ -68,7 +68,7 @@ policyset.caLogSigningSet.8.default.name=Subject Key Identifier Extension Defaul
 policyset.caLogSigningSet.8.default.params.critical=false
 policyset.caLogSigningSet.9.constraint.class_id=signingAlgConstraintImpl
 policyset.caLogSigningSet.9.constraint.name=No Constraint
-policyset.caLogSigningSet.9.constraint.params.signingAlgsAllowed=SHA1withRSA,SHA256withRSA,SHA512withRSA,SHA1withDSA,SHA1withEC,SHA256withEC,SHA384withRSA,SHA384withEC,SHA512withEC,SHA256withRSA/PSS,SHA384withRSA/PSS,SHA512withRSA/PSS
+policyset.caLogSigningSet.9.constraint.params.signingAlgsAllowed=SHA256withRSA,SHA512withRSA,SHA256withEC,SHA384withRSA,SHA384withEC,SHA512withEC,SHA256withRSA/PSS,SHA384withRSA/PSS,SHA512withRSA/PSS
 policyset.caLogSigningSet.9.default.class_id=signingAlgDefaultImpl
 policyset.caLogSigningSet.9.default.name=Signing Alg
 policyset.caLogSigningSet.9.default.params.signingAlg=-
